FailedFiber

+++ resource/closable_io.go
@@ -14,7 +14,7 @@ func FromClosableIO[A ClosableIO](ioa io.IO[A]) Resource[A] {
 }
 
 // BoundedExecutionContextResource returns a resource that is a bounded execution context.
-func BoundedExecutionContextResource(size int64, queueLimit int) Resource[io.ExecutionContext] {
+func BoundedExecutionContextResource(size int, queueLimit int) Resource[io.ExecutionContext] {
 	return FromClosableIO(io.Pure(func() io.ExecutionContext {
 		return io.BoundedExecutionContext(size, queueLimit)
 	}))
